+++ CHANGELOG.md
@@ -10,6 +10,8 @@ _No added features._
 
 ### Changed/Fixed Features
 
+- Updated the confirmation and success messages in the `taskkill` command to display the process name instead of the PID when using the PID.
+- Fixed an issue where the `taskkill` command would detect if the process did not exist after the confirmation prompt was accepted when using PIDs.
 - Fixed an issue where the configuration file version checker would fail to detect an upgrade in some cases.
 
 ### Removed Features